Extract config-modal from dashboard-main.js
